Estabelece save nativo e fundacao multirregional

+++ src/battle_anim_mons.c
@@ -1908,7 +1908,7 @@ void StorePointerInVars(s16 *lo, s16 *hi, const void *ptr)
 
 void *LoadPointerFromVars(s16 lo, s16 hi)
 {
-    return (void *)((u16)lo | ((u16)hi << 16));
+    return (void *)((u16)lo | ((u32)(u16)hi << 16));
 }
 
 void PrepareEruptAnimTaskData(struct Task *task, u8 spriteId, s16 xScaleStart, s16 yScaleStart, s16 xScaleEnd, s16 yScaleEnd, u16 duration)
